# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AgroNutri AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /farms:
    get:
      operationId: listFarms
      tags: [farms]
      responses:
        "200":
          description: Farms visible to the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Farm"
    post:
      operationId: createFarm
      tags: [farms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
  /farms/{farmId}:
    get:
      operationId: getFarm
      tags: [farms]
      parameters:
        - $ref: "#/components/parameters/FarmId"
      responses:
        "200":
          description: Farm detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
    patch:
      operationId: updateFarm
      tags: [farms]
      parameters:
        - $ref: "#/components/parameters/FarmId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
    delete:
      operationId: deleteFarm
      tags: [farms]
      parameters:
        - $ref: "#/components/parameters/FarmId"
      responses:
        "204":
          description: Deleted
  /farms/{farmId}/summary:
    get:
      operationId: getFarmSummary
      tags: [farms]
      parameters:
        - $ref: "#/components/parameters/FarmId"
      responses:
        "200":
          description: Farm context panel summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmSummary"
  /farms/{farmId}/sectors:
    get:
      operationId: listSectors
      tags: [sectors]
      parameters:
        - $ref: "#/components/parameters/FarmId"
      responses:
        "200":
          description: Sectors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sector"
    post:
      operationId: createSector
      tags: [sectors]
      parameters:
        - $ref: "#/components/parameters/FarmId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sector"
  /farms/{farmId}/sectors/{sectorId}:
    patch:
      operationId: updateSector
      tags: [sectors]
      parameters:
        - $ref: "#/components/parameters/FarmId"
        - $ref: "#/components/parameters/SectorId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sector"
    delete:
      operationId: deleteSector
      tags: [sectors]
      parameters:
        - $ref: "#/components/parameters/FarmId"
        - $ref: "#/components/parameters/SectorId"
      responses:
        "204":
          description: Deleted
  /farms/{farmId}/members:
    get:
      operationId: listMembers
      tags: [members]
      parameters:
        - $ref: "#/components/parameters/FarmId"
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FarmMember"
    post:
      operationId: addMember
      tags: [members]
      parameters:
        - $ref: "#/components/parameters/FarmId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmMemberInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmMember"
  /farms/{farmId}/members/{memberId}:
    patch:
      operationId: updateMember
      tags: [members]
      parameters:
        - $ref: "#/components/parameters/FarmId"
        - $ref: "#/components/parameters/MemberId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmMemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmMember"
    delete:
      operationId: removeMember
      tags: [members]
      parameters:
        - $ref: "#/components/parameters/FarmId"
        - $ref: "#/components/parameters/MemberId"
      responses:
        "204":
          description: Removed
...
  /farms/{farmId}/api-config:
    get:
      operationId: getFarmApiConfig
      tags: [settings]
      parameters:
        - $ref: "#/components/parameters/FarmId"
      responses:
        "200":
          description: Farm-specific credential override
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmApiConfig"
    put:
      operationId: setFarmApiConfig
      tags: [settings]
      parameters:
        - $ref: "#/components/parameters/FarmId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmApiConfigInput"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmApiConfig"
...
components:
  parameters:
    FarmId:
      name: farmId
      in: path
      required: true
      schema:
        type: integer
    SectorId:
      name: sectorId
      in: path
      required: true
      schema:
        type: integer
    MemberId:
      name: memberId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Farm:
      type: object
      required: [id, ownerId, name, myRole]
      properties:
        id: { type: integer }
        ownerId: { type: integer }
        name: { type: string }
        companyName: { type: ["string", "null"] }
        cif: { type: ["string", "null"] }
        island: { type: ["string", "null"] }
        municipality: { type: ["string", "null"] }
        latitude: { type: ["number", "null"] }
        longitude: { type: ["number", "null"] }
        altitudeM: { type: ["number", "null"] }
        surfaceHa: { type: ["number", "null"] }
        mainCrop: { type: ["string", "null"] }
        variety: { type: ["string", "null"] }
        plantCount: { type: ["integer", "null"] }
        phenologicalStage: { type: ["string", "null"] }
        cropSystem: { type: ["string", "null"] }
        soilType: { type: ["string", "null"] }
        hasDrainage: { type: ["boolean", "null"] }
        foliarAllowed: { type: ["boolean", "null"] }
        hasDesalinatedWater: { type: ["boolean", "null"] }
        desalinatedWaterPct: { type: ["number", "null"] }
        weeklyLitresPerPlant: { type: ["number", "null"] }
        maxEcDsM: { type: ["number", "null"] }
        managementNotes: { type: ["string", "null"] }
        responsibleTechnician: { type: ["string", "null"] }
        myRole: { type: string, description: "owner | technician | manager | viewer" }
        sectorCount: { type: ["integer", "null"] }
    FarmInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        companyName: { type: string }
        cif: { type: string }
        island: { type: string }
        municipality: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        altitudeM: { type: number }
        surfaceHa: { type: number }
        mainCrop: { type: string }
        variety: { type: string }
        plantCount: { type: integer }
        phenologicalStage: { type: string }
        cropSystem: { type: string }
        soilType: { type: string }
        hasDrainage: { type: boolean }
        foliarAllowed: { type: boolean }
        hasDesalinatedWater: { type: boolean }
        desalinatedWaterPct: { type: number }
        weeklyLitresPerPlant: { type: number }
        maxEcDsM: { type: number }
        managementNotes: { type: string }
        responsibleTechnician: { type: string }
    FarmUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        companyName: { type: string }
        cif: { type: string }
        island: { type: string }
        municipality: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        altitudeM: { type: number }
        surfaceHa: { type: number }
        mainCrop: { type: string }
        variety: { type: string }
        plantCount: { type: integer }
        phenologicalStage: { type: string }
        cropSystem: { type: string }
        soilType: { type: string }
        hasDrainage: { type: boolean }
        foliarAllowed: { type: boolean }
        hasDesalinatedWater: { type: boolean }
        desalinatedWaterPct: { type: number }
        weeklyLitresPerPlant: { type: number }
        maxEcDsM: { type: number }
        managementNotes: { type: string }
        responsibleTechnician: { type: string }
    FarmSummary:
      type: object
      required: [farm]
      properties:
        farm:
          $ref: "#/components/schemas/Farm"
        weeklyWaterM3: { type: ["number", "null"] }
        latestSoilAnalysis:
          oneOf:
            - $ref: "#/components/schemas/Analysis"
            - type: "null"
        latestLeafAnalysis:
          oneOf:
            - $ref: "#/components/schemas/Analysis"
            - type: "null"
        latestWaterAnalysis:
          oneOf:
            - $ref: "#/components/schemas/Analysis"
            - type: "null"
        activeRecommendation:
          oneOf:
            - $ref: "#/components/schemas/Recommendation"
            - type: "null"
        alerts:
          type: array
          items:
            type: string
        aiAvailable: { type: boolean }
    Sector:
      type: object
      required: [id, farmId, name]
      properties:
        id: { type: integer }
        farmId: { type: integer }
        name: { type: string }
        plantCount: { type: ["integer", "null"] }
        surfaceHa: { type: ["number", "null"] }
        weeklyLitresPerPlant: { type: ["number", "null"] }
        phenologicalStage: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
    SectorInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        plantCount: { type: integer }
        surfaceHa: { type: number }
        weeklyLitresPerPlant: { type: number }
        phenologicalStage: { type: string }
        notes: { type: string }
    SectorUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        plantCount: { type: integer }
        surfaceHa: { type: number }
        weeklyLitresPerPlant: { type: number }
        phenologicalStage: { type: string }
        notes: { type: string }
    FarmMember:
      type: object
      required: [id, farmId, userId, role, name, email]
      properties:
        id: { type: integer }
        farmId: { type: integer }
        userId: { type: integer }
        role: { type: string, description: "technician | manager | viewer" }
        name: { type: string }
        email: { type: string }
    FarmMemberInput:
      type: object
      required: [email, role]
      properties:
        email: { type: string, format: email }
        role: { type: string, enum: [technician, manager, viewer] }
    FarmMemberUpdate:
      type: object
      required: [role]
      properties:
        role: { type: string, enum: [technician, manager, viewer] }
    Analysis:
      type: object
      required: [id, farmId, type, sampleDate, parameters]
      properties:
        id: { type: integer }
        farmId: { type: integer }
        sectorId: { type: ["integer", "null"] }
        type: { type: string, enum: [soil, leaf, water] }
        reference: { type: ["string", "null"] }
        laboratory: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        sampleDate: { type: string, description: "YYYY-MM-DD" }
        parameters:
          type: array
          items:
            $ref: "#/components/schemas/AnalysisParameter"
        notes: { type: ["string", "null"] }
        createdAt: { type: ["string", "null"] }
    AnalysisParameter:
      type: object
      required: [name, value]
      properties:
        name: { type: string }
        value: { type: number }
        unit: { type: ["string", "null"] }
        refLow: { type: ["number", "null"] }
        refHigh: { type: ["number", "null"] }
        status: { type: ["string", "null"], description: "muy_bajo | bajo | normal | alto | muy_alto" }
...
    RecommendationItem:
      type: object
      required: [fertilizerName, weeklyDose, unit]
      properties:
        fertilizerId: { type: ["integer", "null"] }
        fertilizerName: { type: string }
        weeklyDose: { type: number }
        unit: { type: string, description: "kg | L" }
        previousDose: { type: ["number", "null"] }
        reason: { type: ["string", "null"] }
    Recommendation:
      type: object
      required: [id, farmId, status, items, createdAt]
      properties:
        id: { type: integer }
        farmId: { type: integer }
        sectorId: { type: ["integer", "null"] }
        title: { type: ["string", "null"] }
        status: { type: string, description: "draft | pending_review | validated | applying | finished | rejected" }
        source: { type: ["string", "null"], description: "manual | ai" }
        items:
          type: array
          items:
            $ref: "#/components/schemas/RecommendationItem"
        rationale: { type: ["string", "null"] }
        estimatedEcDsM: { type: ["number", "null"] }
        estimatedWeeklyNKg: { type: ["number", "null"] }
        warnings:
          type: array
          items:
            type: string
        createdByName: { type: ["string", "null"] }
        validatedByName: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: ["string", "null"] }
...
    FarmApiConfig:
      type: object
      required: [farmId]
      properties:
        farmId: { type: integer }
        credentialId: { type: ["integer", "null"] }
        credentialName: { type: ["string", "null"] }
    FarmApiConfigInput:
      type: object
      properties:
        credentialId: { type: ["integer", "null"], description: "null = use user's default credential" }
...